fixing some -Wsign-compare stuff

+++ libraries/threads/threads.c
@@ -1,4 +1,4 @@
-// Copyright (C) 2020 Björn Lindqvist <bjourne@gmail.com>
+// Copyright (C) 2020, 2023 Björn A. Lindqvist <bjourne@gmail.com>
 #include <inttypes.h>
 #include <stdio.h>
 #include "threads/threads.h"
@@ -7,7 +7,7 @@ bool
 thr_create_threads(size_t n, thr_handle *handles,
                    size_t size, void *args,
                    void *(*func) (void *)) {
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         void *arg = (void *)((char*)args + size * i);
 #if _WIN32
         handles[i] = CreateThread(NULL, 0, (LPTHREAD_START_ROUTINE)func,
@@ -26,7 +26,7 @@ thr_create_threads(size_t n, thr_handle *handles,
 
 bool
 thr_wait_for_threads(size_t n, thr_handle *handles) {
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
 #if _WIN32
         WaitForSingleObject(handles[i], INFINITE);
 #else
